+++ Rules/IPv4/Inbound/StoreApps.ps1
@@ -137,7 +137,7 @@ foreach ($Principal in $Principals)
 			return
 		}
 
-		$PackageSID = Get-AppSID $Principal.User $_.PackageFamilyName
+		$PackageSID = Get-AppSID $_.PackageFamilyName
 
 		# Possible package not found
 		if ($PackageSID)
@@ -162,7 +162,7 @@ foreach ($Principal in $Principals)
 	# Create rules for system apps
 	#
 
-	Get-SystemApps | ForEach-Object -Process {
+	Get-SystemApps -User $Principal.User | ForEach-Object -Process {
 		$NetworkCapabilities = $_ | Get-AppCapability -Networking
 
 		if (!$NetworkCapabilities)
@@ -194,7 +194,7 @@ foreach ($Principal in $Principals)
 			return
 		}
 
-		$PackageSID = Get-AppSID $Principal.User $_.PackageFamilyName
+		$PackageSID = Get-AppSID $_.PackageFamilyName
 
 		# Possible package not found
 		if ($PackageSID)